Add autoConfigure option so user can disable broker autoconfiguration
To disable auto configuration of the broker, set system property `-Dserver.auto-configure=false`
or environment variable `SERVER_AUTO_CONFIGURE=false`. This should be used in combination
with `SERVER_PROPERTIES_FILE`.

Why:
Users may wish to supply all the configuration themselves to test odd scenarios
like a node in controller-only mode, where advertised listeners cannot be set.
Configurations like this may be outside of the scope of what the auto configuration
is trying to cover. In this case we can provide an escape hatch for users to supply
all the configuration.